pyproject.toml
[build-system]
requires = ['setuptools>=60', 'setuptools_scm>=8.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'mccode-plumber'
dependencies = [
    'p4p',
    'kafka-python>=2.2.11',
    'ess-streaming-data-types>=0.14.0',
    'restage>=0.10.0',
    'mccode-to-kafka>=0.3.1',
    'moreniius>=0.6.2',
    'icecream',
    'ephemeral-port-reserve',
]
readme = "README.md"
authors = [
    { name = "Gregory Tucker", email = "gregory.tucker@ess.eu" },
]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dynamic = ['version']

[project.optional-dependencies]
test = ["pytest", "niess>=0.1.4"]

[project.scripts]
mp-splitrun = 'mccode_plumber.splitrun:main'
mp-epics = 'mccode_plumber.epics:run'
mp-epics-strings = 'mccode_plumber.epics:run_strings'
mp-epics-update = 'mccode_plumber.epics:update'
mp-epics-watch = 'mccode_plumber.epics_watcher:run_instr'
mp-forwarder-setup = 'mccode_plumber.forwarder:setup'
mp-forwarder-teardown = 'mccode_plumber.forwarder:teardown'
mp-writer-from = 'mccode_plumber.writer:print_time'
mp-writer-write = 'mccode_plumber.writer:start_writer'
mp-writer-wait = 'mccode_plumber.writer:wait_on_writer'
mp-writer-list = 'mccode_plumber.writer:list_status'
mp-writer-kill = 'mccode_plumber.writer:kill_job'
mp-writer-killall = 'mccode_plumber.writer:kill_all'
mp-register-topics = 'mccode_plumber.kafka:register_topics'
mp-insert-hdf5-instr = 'mccode_plumber.mccode:insert'
mp-nexus-splitrun = 'mccode_plumber.manage.orchestrate:main'
mp-nexus-services = 'mccode_plumber.manage.orchestrate:services'

[tool.setuptools_scm]

[tool.tox]
legacy_tox_ini = """
    [tox]
    min_version = 4.0
    env_list =
        py313
        py312
        py311
        type

    [testenv]
    deps =
        pytest
        niess>=0.1.4
    commands = pytest tests

    [testenv:type]
    deps =
        mypy
        types-PyYAML
    commands = mypy src
"""
